+++ tests/test_class_schema.py
@@ -1,6 +1,7 @@
+import inspect
 import typing
 import unittest
-from typing import Any, TYPE_CHECKING
+from typing import Any, cast, TYPE_CHECKING
 from uuid import UUID
 
 try:
@@ -38,12 +39,60 @@ class TestClassSchema(unittest.TestCase):
         complex_set = {
             class_schema(ComplexNested),
             class_schema(ComplexNested, base_schema=None),
+            class_schema(ComplexNested, clazz_frame=None),
             class_schema(ComplexNested, None),
+            class_schema(ComplexNested, None, None),
         }
         simple_set = {
             class_schema(Simple),
             class_schema(Simple, base_schema=None),
+            class_schema(Simple, clazz_frame=None),
             class_schema(Simple, None),
+            class_schema(Simple, None, None),
+        }
+        self.assertEqual(len(complex_set), 1)
+        self.assertEqual(len(simple_set), 1)
+
+    def test_nested_schema_with_passed_frame(self):
+        @dataclasses.dataclass
+        class Simple:
+            one: str = dataclasses.field()
+            two: str = dataclasses.field()
+
+        @dataclasses.dataclass
+        class ComplexNested:
+            three: int = dataclasses.field()
+            four: Simple = dataclasses.field()
+
+        frame = inspect.stack()[0][0]
+
+        self.assertIs(
+            class_schema(ComplexNested, clazz_frame=frame),
+            class_schema(ComplexNested, clazz_frame=frame),
+        )
+        self.assertIs(
+            class_schema(Simple, clazz_frame=frame),
+            class_schema(Simple, clazz_frame=frame),
+        )
+        self.assertIs(
+            class_schema(Simple, clazz_frame=frame),
+            class_schema(ComplexNested, clazz_frame=frame)
+            ._declared_fields["four"]
+            .nested,
+        )
+
+        complex_set = {
+            class_schema(ComplexNested, clazz_frame=frame),
+            class_schema(ComplexNested, base_schema=None, clazz_frame=frame),
+            class_schema(ComplexNested, None, clazz_frame=frame),
+            class_schema(ComplexNested, None, frame),
+        }
+        simple_set = {
+            class_schema(Simple, clazz_frame=frame),
+            class_schema(Simple, base_schema=None, clazz_frame=frame),
+            class_schema(Simple, None, clazz_frame=frame),
+            class_schema(Simple, clazz_frame=frame),
+            class_schema(Simple, None, frame),
         }
         self.assertEqual(len(complex_set), 1)
         self.assertEqual(len(simple_set), 1)
@@ -178,7 +227,7 @@ class TestClassSchema(unittest.TestCase):
                 schema.load({"data": data})
 
     def test_final_infers_type_from_default(self):
-        # @dataclasses.dataclass
+        # @dataclasses.dataclass(frozen=True)
         class A:
             data: Final = "a"
 
@@ -189,10 +238,12 @@ class TestClassSchema(unittest.TestCase):
         # NOTE: This workaround is needed to avoid a Mypy crash.
         # See: https://github.com/python/mypy/issues/10090#issuecomment-865971891
         if not TYPE_CHECKING:
-            A = dataclasses.dataclass(A)
+            frozen_dataclass = dataclasses.dataclass(frozen=True)
+            A = frozen_dataclass(A)
             B = dataclasses.dataclass(B)
 
-        schema_a = class_schema(A)()
+        with self.assertWarns(Warning):
+            schema_a = class_schema(A)()
         self.assertEqual(A(data="a"), schema_a.load({}))
         self.assertEqual(A(data="a"), schema_a.load({"data": "a"}))
         self.assertEqual(A(data="b"), schema_a.load({"data": "b"}))
@@ -203,7 +254,8 @@ class TestClassSchema(unittest.TestCase):
             with self.assertRaises(ValidationError):
                 schema_a.load({"data": data})
 
-        schema_b = class_schema(B)()
+        with self.assertWarns(Warning):
+            schema_b = class_schema(B)()
         self.assertEqual(B(data=A()), schema_b.load({}))
         self.assertEqual(B(data=A()), schema_b.load({"data": {}}))
         self.assertEqual(B(data=A()), schema_b.load({"data": {"data": "a"}}))
@@ -215,6 +267,31 @@ class TestClassSchema(unittest.TestCase):
         for data in [2, 2.34, False]:
             with self.assertRaises(ValidationError):
                 schema_b.load({"data": data})
+
+    def test_final_infers_type_any_from_field_default_factory(self):
+        # @dataclasses.dataclass
+        class A:
+            data: Final = dataclasses.field(default_factory=lambda: [])
+
+        # NOTE: This workaround is needed to avoid a Mypy crash.
+        # See: https://github.com/python/mypy/issues/10090#issuecomment-866686096
+        if not TYPE_CHECKING:
+            A = dataclasses.dataclass(A)
+
+        with self.assertWarns(Warning):
+            schema = class_schema(A)()
+
+        self.assertEqual(A(data=[]), schema.load({}))
+        self.assertEqual(A(data=[]), schema.load({"data": []}))
+        self.assertEqual(A(data=["a"]), schema.load({"data": ["a"]}))
+        self.assertEqual(schema.dump(A()), {"data": []})
+        self.assertEqual(schema.dump(A(data=[])), {"data": []})
+        self.assertEqual(schema.dump(A(data=["a"])), {"data": ["a"]})
+
+        # The following test cases pass because the field type cannot be
+        # inferred from a factory.
+        self.assertEqual(A(data=cast(Any, True)), schema.load({"data": True}))
+        self.assertEqual(A(data=cast(Any, "a")), schema.load({"data": "a"}))
 
     def test_validator_stacking(self):
         # See: https://github.com/lovasoa/marshmallow_dataclass/issues/91
@@ -351,6 +428,38 @@ class TestClassSchema(unittest.TestCase):
         with self.assertRaises(ValidationError):
             schema_n.load({"data": {"data": "str"}})
 
+    def test_recursive_reference(self):
+        @dataclasses.dataclass
+        class Tree:
+            children: typing.List["Tree"]  # noqa: F821
+
+        schema = class_schema(Tree)()
+
+        self.assertEqual(
+            schema.load({"children": [{"children": []}]}),
+            Tree(children=[Tree(children=[])]),
+        )
+
+    def test_cyclic_reference(self):
+        @dataclasses.dataclass
+        class First:
+            second: typing.Optional["Second"]  # noqa: F821
+
+        @dataclasses.dataclass
+        class Second:
+            first: typing.Optional["First"]
+
+        first_schema = class_schema(First)()
+        second_schema = class_schema(Second)()
+
+        self.assertEqual(
+            first_schema.load({"second": {"first": None}}),
+            First(second=Second(first=None)),
+        )
+        self.assertEqual(
+            second_schema.dump(Second(first=First(second=Second(first=None)))),
+            {"first": {"second": {"first": None}}},
+        )
 
 if __name__ == "__main__":
     unittest.main()