+++ tests/test_field_for_schema.py
@@ -1,4 +1,5 @@
 import inspect
+import sys
 import typing
 import unittest
 from enum import Enum
@@ -53,12 +54,37 @@ class TestFieldForSchema(unittest.TestCase):
             ),
         )
 
+    def test_dict_from_typing_wo_args(self):
+        self.assertFieldsEqual(
+            field_for_schema(Dict),
+            fields.Dict(
+                keys=fields.Raw(required=True, allow_none=True),
+                values=fields.Raw(required=True, allow_none=True),
+                required=True,
+            ),
+        )
+
     def test_builtin_dict(self):
         self.assertFieldsEqual(
             field_for_schema(dict),
             fields.Dict(
                 keys=fields.Raw(required=True, allow_none=True),
                 values=fields.Raw(required=True, allow_none=True),
+                required=True,
+            ),
+        )
+
+    def test_list_from_typing(self):
+        self.assertFieldsEqual(
+            field_for_schema(List[int]),
+            fields.List(fields.Integer(required=True), required=True),
+        )
+
+    def test_list_from_typing_wo_args(self):
+        self.assertFieldsEqual(
+            field_for_schema(List),
+            fields.List(
+                fields.Raw(required=True, allow_none=True),
                 required=True,
             ),
         )
@@ -217,6 +243,13 @@ class TestFieldForSchema(unittest.TestCase):
         )
 
     def test_sequence(self):
+        self.maxDiff = 2000
+        self.assertFieldsEqual(
+            field_for_schema(typing.Sequence[int]),
+            collection_field.Sequence(fields.Integer(required=True), required=True),
+        )
+
+    def test_sequence_wo_args(self):
         self.assertFieldsEqual(
             field_for_schema(typing.Sequence),
             collection_field.Sequence(
@@ -225,7 +258,30 @@ class TestFieldForSchema(unittest.TestCase):
             ),
         )
 
-    def test_set(self):
+    def test_homogeneous_tuple_from_typing(self):
+        self.assertFieldsEqual(
+            field_for_schema(Tuple[str, ...]),
+            collection_field.Sequence(fields.String(required=True), required=True),
+        )
+
+    @unittest.skipIf(sys.version_info < (3, 9), "PEP 585 unsupported")
+    def test_homogeneous_tuple(self):
+        self.assertFieldsEqual(
+            field_for_schema(tuple[float, ...]),
+            collection_field.Sequence(fields.Float(required=True), required=True),
+        )
+
+    def test_set_from_typing(self):
+        self.assertFieldsEqual(
+            field_for_schema(typing.Set[str]),
+            collection_field.Set(
+                fields.String(required=True),
+                frozen=False,
+                required=True,
+            ),
+        )
+
+    def test_set_from_typing_wo_args(self):
         self.assertFieldsEqual(
             field_for_schema(typing.Set),
             collection_field.Set(
@@ -235,9 +291,39 @@ class TestFieldForSchema(unittest.TestCase):
             ),
         )
 
-    def test_frozenset(self):
+    def test_builtin_set(self):
+        self.assertFieldsEqual(
+            field_for_schema(set),
+            collection_field.Set(
+                cls_or_instance=fields.Raw(required=True, allow_none=True),
+                frozen=False,
+                required=True,
+            ),
+        )
+
+    def test_frozenset_from_typing(self):
+        self.assertFieldsEqual(
+            field_for_schema(typing.FrozenSet[int]),
+            collection_field.Set(
+                fields.Integer(required=True),
+                frozen=True,
+                required=True,
+            ),
+        )
+
+    def test_frozenset_from_typing_wo_args(self):
         self.assertFieldsEqual(
             field_for_schema(typing.FrozenSet),
+            collection_field.Set(
+                cls_or_instance=fields.Raw(required=True, allow_none=True),
+                frozen=True,
+                required=True,
+            ),
+        )
+
+    def test_builtin_frozenset(self):
+        self.assertFieldsEqual(
+            field_for_schema(frozenset),
             collection_field.Set(
                 cls_or_instance=fields.Raw(required=True, allow_none=True),
                 frozen=True,